+++ fixtures/flv/index.js
@@ -15,7 +15,9 @@ function defaultOpt() {
     bufferBehind: 10,
     maxJumpDistance: 3,
     maxReaderInterval: 5000,
-    seamlesslyReload: false
+    seamlesslyReload: false,
+    firstMaxChunkSize: 20000,
+    manualLoad: true
   }
 }
 var cachedOpt = localStorage.getItem('xg:test:flv:opt')
@@ -29,6 +31,9 @@ var testPoint = Number(localStorage.getItem('xg:test:flv:point'))
 if (isNaN(testPoint)) testPoint = 0
 
 window.onload = function () {
+  fetch('https://pull-demo.volcfcdnrd.com/live/st-4536524.flv').then(res => {
+    window.streamRes = res
+  })
   var dTestPoint = document.getElementById('test-point')
   var dTestPointDesc = document.getElementById('test-point-desc')
 
@@ -105,6 +110,7 @@ window.onload = function () {
       init()
     }
     function init() {
+      window.timeStart = Date.now()
       window.player = player = new Player({
         el: document.getElementById('player'),
         plugins: [FlvPlayer],
@@ -112,12 +118,26 @@ window.onload = function () {
         isLive: opts.isLive,
         autoplay: opts.autoplay,
         autoplayMuted: opts.autoplayMuted,
-        flv: opts
+        flv: {
+          // streamRes: window.streamRes,
+          ...opts
+        }
       });
+      player.once('ready', () => {
+        // console.log('streamRes', Date.now() - timeStart, streamRes, player.plugins.flv)
+        if (player.config.flv.manualLoad) {
+          player.plugins.flv.loadSource(player.config.url, streamRes)
+          streamRes = null
+        }
+      })
       dlEvent.innerHTML = ''
       dlError.innerHTML = ''
 
       function pushEvent(name, value, container) {
+        if (name === 'loadeddata') {
+          console.log('loadeddata', Date.now() - window.timeStart)
+          player.plugins.flv.flv._mediaLoader._currentTask._loader._firstMaxChunkSize = null
+        }
         container = container || dlEvent
         if (container === dlEvent && dlLogPause.checked) return
         // console.debug('[test]', name, value)
@@ -252,7 +272,9 @@ window.onload = function () {
   inp(doDisconnectTime).onchange = function () { updateOpts('disconnectTime', this.value, 'number') }
   inp(doMaxReaderInterval).onchange = function () { updateOpts('maxReaderInterval', this.value, 'number') }
 
-  initPlayer()
+  setTimeout(() => {
+    initPlayer()
+  }, 3000)
 
   dbResetOpt.onclick = resetOpts
   dbApplyOpt.onclick = initPlayer