+++ packages/xgplayer/src/index.umd.js
@@ -1,16 +1,21 @@
-import Player from './player'
-import Plugin from './plugin/plugin'
-import BasePlugin from './plugin/basePlugin'
-import * as Events from './events'
-import STATE_CLASS from './stateClassMap'
-import I18N from './lang/i18n'
 import Errors from './error'
+import * as Events from './events'
+import './index.scss'
+import { InstManager } from './instManager'
+import I18N from './lang/i18n'
+import Player from './player'
+import BasePlugin from './plugin/basePlugin'
+import Plugin from './plugin/plugin'
+import { default as CssFullscreenIcon } from './plugins/cssFullScreen'
+import { default as FullscreenIcon } from './plugins/fullscreen'
+import { default as PlayIcon } from './plugins/play'
+import { default as Progress } from './plugins/progress'
+import { default as TimeIcon } from './plugins/time'
+import { default as VolumeIcon } from './plugins/volume'
+import defaultPreset from './presets/default'
+import STATE_CLASS from './stateClassMap'
 import Sniffer from './utils/sniffer'
 import Util from './utils/util'
-import { InstManager } from './instManager'
-import './index.scss'
-
-import defaultPreset from './presets/default'
 
 class PresetPlayer extends Player {
   static defaultPreset = defaultPreset;
@@ -23,6 +28,14 @@ class PresetPlayer extends Player {
   static I18N = I18N
   static STATE_CLASS = STATE_CLASS
   static InstManager = InstManager
+
+  // UI Exports
+  static PlayIcon = PlayIcon // For Ads Plugin
+  static TimeIcon = TimeIcon // For Ads Plugin
+  static Progress = Progress // For Ads Plugin
+  static FullscreenIcon = FullscreenIcon // For Ads Plugin
+  static CssFullscreenIcon = CssFullscreenIcon // For Ads Plugin
+  static VolumeIcon = VolumeIcon // For Ads Plugin
 }
 
 export default PresetPlayer