+++ packages/xgplayer/src/plugins/progress/index.js
@@ -472,6 +472,9 @@ class Progress extends Plugin {
 
   onMouseUp = (e) => {
     const { player, config, pos, playerConfig, _state } = this
+    if (!pos) {
+      return
+    }
     e.stopPropagation()
     e.preventDefault()
     Util.checkIsFunction(playerConfig.enableSwipeHandler) && playerConfig.enableSwipeHandler()
@@ -620,6 +623,9 @@ class Progress extends Plugin {
     offset = offset > rWidth ? rWidth : (offset < 0 ? 0 : offset)
     let percent = offset / rWidth
     percent = percent < 0 ? 0 : (percent > 1 ? 1 : percent)
+    if(Number.isNaN(percent)){
+      percent = this.player.currentTime / this.offsetDuration
+    }
     const currentTime = parseInt(percent * this.offsetDuration * 1000, 10) / 1000
     const seekTime = Util.getCurrentTimeByOffset(currentTime, player.timeSegments)
     return {